+++ src/babyCheckValid.ts
@@ -26,7 +26,7 @@ import {
 } from "isaacscript-common";
 import type { Baby } from "./classes/Baby";
 import {
-  BLINDFOLDED_ANTI_SYNERGY_COLLECTIBLES,
+  BLINDFOLDED_ANTI_SYNERGY_COLLECTIBLE_TYPES,
   BRIMSTONE_ANTI_SYNERGIES,
   COLLECTIBLES_THAT_REMOVE_TEARS,
   C_SECTION_ANTI_SYNERGIES,
@@ -118,7 +118,7 @@ function checkCollectibles(
 
   if (
     baby.blindfolded === true &&
-    !hasCollectible(player, ...BLINDFOLDED_ANTI_SYNERGY_COLLECTIBLES)
+    !hasCollectible(player, ...BLINDFOLDED_ANTI_SYNERGY_COLLECTIBLE_TYPES)
   ) {
     return false;
   }